Change Class Name & Add DataType Class Instance
